browser/components/payments/test/mochitest/test_address_form.html

<!DOCTYPE HTML>
<html>
<!--
Test the address-form element
-->
<head>
  <meta charset="utf-8">
  <title>Test the address-form element</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <script src="sinon-2.3.2.js"></script>
  <script src="payments_common.js"></script>
  <script src="../../res/vendor/custom-elements.min.js"></script>
  <script src="../../res/unprivileged-fallbacks.js"></script>
  <script src="autofillEditForms.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <link rel="stylesheet" type="text/css" href="../../res/paymentRequest.css"/>
  <link rel="stylesheet" type="text/css" href="../../res/containers/address-form.css"/>
</head>
<body>
  <p id="display">
  </p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<script type="module">
/** Test the address-form element **/

/* global sinon, PaymentDialogUtils */

import AddressForm from "../../res/containers/address-form.js";

let display = document.getElementById("display");

function checkAddressForm(customEl, expectedAddress) {
  const ADDRESS_PROPERTY_NAMES = [
    "given-name",
    "family-name",
    "organization",
    "street-address",
    "address-level2",
    "address-level1",
    "postal-code",
    "country",
    "email",
    "tel",
  ];
  for (let propName of ADDRESS_PROPERTY_NAMES) {
    let expectedVal = expectedAddress[propName] || "";
    is(document.getElementById(propName).value,
       expectedVal.toString(),
       `Check ${propName}`);
  }
}

function sendStringAndCheckValidity(element, string, isValid) {
  element.focus();
  while (element.value) {
    sendKey("BACK_SPACE");
  }
  sendString(string);
  ok(element.checkValidity() == isValid,
     `${element.id} should be ${isValid ? "valid" : "invalid"} (${string})`);
}

add_task(async function test_initialState() {
  let form = new AddressForm();
  let {page} = form.requestStore.getState();
  is(page.id, "payment-summary", "Check initial page");
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();
  is(page.id, "payment-summary", "Check initial page after appending");
  form.remove();
});

add_task(async function test_backButton() {
  let form = new AddressForm();
  form.dataset.backButtonLabel = "Back";
  await form.requestStore.setState({
    page: {
      id: "address-page",
    },
    "address-page": {
      title: "Sample page title",
    },
  });
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();

  let stateChangePromise = promiseStateChange(form.requestStore);
  is(form.pageTitle.textContent, "Sample page title", "Check label");

  is(form.backButton.textContent, "Back", "Check label");
  form.backButton.scrollIntoView();
  synthesizeMouseAtCenter(form.backButton, {});

  let {page} = await stateChangePromise;
  is(page.id, "payment-summary", "Check initial page after appending");

  form.remove();
});

add_task(async function test_saveButton() {
  let form = new AddressForm();
  form.dataset.saveButtonLabel = "Save";
  form.dataset.errorGenericSave = "Generic error";
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();

  form.form.querySelector("#given-name").focus();
  sendString("Jaws");
  form.form.querySelector("#family-name").focus();
  sendString("Swaj");
  form.form.querySelector("#organization").focus();
  sendString("Allizom");
  form.form.querySelector("#street-address").focus();
  sendString("404 Internet Super Highway");
  form.form.querySelector("#address-level2").focus();
  sendString("Firefoxity City");
  form.form.querySelector("#address-level1").focus();
  sendString("CA");
  form.form.querySelector("#postal-code").focus();
  sendString("00001");
  form.form.querySelector("#country option[value='US']").selected = true;
  form.form.querySelector("#email").focus();
  sendString("test@example.com");
  form.form.querySelector("#tel").focus();
  sendString("+15555551212");

  let messagePromise = promiseContentToChromeMessage("updateAutofillRecord");
  is(form.saveButton.textContent, "Save", "Check label");
  form.saveButton.scrollIntoView();
  synthesizeMouseAtCenter(form.saveButton, {});

  let details = await messagePromise;
  is(details.collectionName, "addresses", "Check collectionName");
  isDeeply(details, {
    collectionName: "addresses",
    errorStateChange: {
      page: {
        id: "address-page",
        error: "Generic error",
        onboardingWizard: undefined,
      },
      "address-page": {
        title: "Sample page title",
      },
    },
    guid: undefined,
    messageType: "updateAutofillRecord",
    preserveOldProperties: true,
    record: {
      "given-name": "Jaws",
      "family-name": "Swaj",
      "organization": "Allizom",
      "street-address": "404 Internet Super Highway",
      "address-level2": "Firefoxity City",
      "address-level1": "CA",
      "postal-code": "00001",
      "country": "US",
      "email": "test@example.com",
      "tel": "+15555551212",
    },
    selectedStateKey: undefined,
    successStateChange: {
      page: {
        id: "payment-summary",
        onboardingWizard: undefined,
      },
    },
  }, "Check event details for the message to chrome");
  form.remove();
});

add_task(async function test_genericError() {
  let form = new AddressForm();
  await form.requestStore.setState({
    page: {
      id: "test-page",
      error: "Generic Error",
    },
  });
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();

  ok(!isHidden(form.genericErrorText), "Error message should be visible");
  is(form.genericErrorText.textContent, "Generic Error", "Check error message");
  form.remove();
});

add_task(async function test_edit() {
  let form = new AddressForm();
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();

  let address1 = deepClone(PTU.Addresses.TimBL);
  address1.guid = "9864798564";

  await form.requestStore.setState({
    page: {
      id: "address-page",
    },
    "address-page": {
      guid: address1.guid,
    },
    savedAddresses: {
      [address1.guid]: deepClone(address1),
    },
  });
  await asyncElementRendered();
  checkAddressForm(form, address1);

  info("test change to minimal record");
  let minimalAddress = {
    "given-name": address1["given-name"],
    guid: "9gnjdhen46",
  };
  await form.requestStore.setState({
    page: {
      id: "address-page",
    },
    "address-page": {
      guid: minimalAddress.guid,
    },
    savedAddresses: {
      [minimalAddress.guid]: deepClone(minimalAddress),
    },
  });
  await asyncElementRendered();
  checkAddressForm(form, minimalAddress);

  info("change to no selected address");
  await form.requestStore.setState({
    page: {
      id: "address-page",
    },
    "address-page": {},
  });
  await asyncElementRendered();
  checkAddressForm(form, {});

  form.remove();
});

add_task(async function test_restricted_address_fields() {
  let form = new AddressForm();
  form.dataset.saveButtonLabel = "Save";
  form.dataset.errorGenericSave = "Generic error";
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();
  form.setAttribute("address-fields", "name email tel");

  ok(!isHidden(form.form.querySelector("#given-name")),
     "given-name should be visible");
  ok(!isHidden(form.form.querySelector("#additional-name")),
     "additional-name should be visible");
  ok(!isHidden(form.form.querySelector("#family-name")),
     "family-name should be visible");
  ok(isHidden(form.form.querySelector("#organization")),
     "organization should be hidden");
  ok(isHidden(form.form.querySelector("#street-address")),
     "street-address should be hidden");
  ok(isHidden(form.form.querySelector("#address-level2")),
     "address-level2 should be hidden");
  ok(isHidden(form.form.querySelector("#address-level1")),
     "address-level1 should be hidden");
  ok(isHidden(form.form.querySelector("#postal-code")),
     "postal-code should be hidden");
  ok(isHidden(form.form.querySelector("#country")),
     "country should be hidden");
  ok(!isHidden(form.form.querySelector("#email")),
     "email should be visible");
  ok(!isHidden(form.form.querySelector("#tel")),
     "tel should be visible");

  form.remove();
});

add_task(async function test_field_validation() {
  let form = new AddressForm();
  await form.promiseReady;
  display.appendChild(form);
  await asyncElementRendered();

  let postalCodeInput = form.form.querySelector("#postal-code");
  let addressLevel1Input = form.form.querySelector("#address-level1");
  ok(!postalCodeInput.value, "postal-code should be empty by default");
  ok(!addressLevel1Input.value, "address-level1 should be empty by default");
  ok(!postalCodeInput.checkValidity(), "postal-code should be invalid by default");
  ok(!addressLevel1Input.checkValidity(), "address-level1 should be invalid by default");

  let countrySelect = form.form.querySelector("#country");
  countrySelect.selectedIndex = [...countrySelect.options].findIndex(o => o.value == "US");
  countrySelect.dispatchEvent(new Event("change"));

  sendStringAndCheckValidity(addressLevel1Input, "MI", true);
  sendStringAndCheckValidity(addressLevel1Input, "", false);
  sendStringAndCheckValidity(postalCodeInput, "B4N4N4", false);
  sendStringAndCheckValidity(addressLevel1Input, "Nova Scotia", true);
  sendStringAndCheckValidity(postalCodeInput, "R3J 3C7", false);
  sendStringAndCheckValidity(addressLevel1Input, "", false);
  sendStringAndCheckValidity(postalCodeInput, "11109", true);
  sendStringAndCheckValidity(addressLevel1Input, "Nova Scotia", true);
  sendStringAndCheckValidity(postalCodeInput, "06390-0001", true);

  countrySelect.selectedIndex = [...countrySelect.options].findIndex(o => o.value == "CA");
  countrySelect.dispatchEvent(new Event("change"));

  sendStringAndCheckValidity(postalCodeInput, "00001", false);
  sendStringAndCheckValidity(addressLevel1Input, "CA", true);
  sendStringAndCheckValidity(postalCodeInput, "94043", false);
  sendStringAndCheckValidity(addressLevel1Input, "", false);
  sendStringAndCheckValidity(postalCodeInput, "B4N4N4", true);
  sendStringAndCheckValidity(addressLevel1Input, "MI", true);
  sendStringAndCheckValidity(postalCodeInput, "R3J 3C7", true);
  sendStringAndCheckValidity(addressLevel1Input, "", false);
  sendStringAndCheckValidity(postalCodeInput, "11109", false);
  sendStringAndCheckValidity(addressLevel1Input, "Nova Scotia", true);
  sendStringAndCheckValidity(postalCodeInput, "06390-0001", false);

  form.remove();
});

add_task(async function test_customValidity() {
  let form = new AddressForm();
  await form.promiseReady;
  const state = {
    page: {
      id: "address-page",
    },
    "address-page": {
      title: "Sample page title",
    },
    request: {
      paymentDetails: {
        shippingAddressErrors: {
          addressLine: "Street address needs to start with a D",
          city: "City needs to start with a B",
          country: "Country needs to start with a C",
          organization: "organization needs to start with an A",
          phone: "Telephone needs to start with a 9",
          postalCode: "Postal code needs to start with a 0",
          recipient: "Name needs to start with a Z",
          region: "Region needs to start with a Y",
        },
      },
    },
  };
  await form.requestStore.setState(state);
  display.appendChild(form);
  await asyncElementRendered();

  function checkValidationMessage(selector, property) {
    is(form.form.querySelector(selector).validationMessage,
       state.request.paymentDetails.shippingAddressErrors[property],
       "Validation message should match for " + selector);
  }

  checkValidationMessage("#street-address", "addressLine");
  checkValidationMessage("#address-level2", "city");
  checkValidationMessage("#country", "country");
  checkValidationMessage("#organization", "organization");
  checkValidationMessage("#tel", "phone");
  checkValidationMessage("#postal-code", "postalCode");
  checkValidationMessage("#given-name", "recipient");
  checkValidationMessage("#address-level1", "region");

  form.remove();
});

add_task(async function test_field_validation() {
  sinon.stub(PaymentDialogUtils, "getFormFormat").returns({
    addressLevel1Label: "state",
    postalCodeLabel: "US",
    fieldsOrder: [
      {fieldId: "name", newLine: true},
      {fieldId: "organization", newLine: true},
      {fieldId: "street-address", newLine: true},
      {fieldId: "address-level2"},
    ],
  });

  let form = new AddressForm();
  await form.promiseReady;
  const state = {
    page: {
      id: "address-page",
    },
    "address-page": {
      title: "Sample page title",
    },
    request: {
      paymentDetails: {
        shippingAddressErrors: {},
      },
    },
  };
  await form.requestStore.setState(state);
  display.appendChild(form);
  await asyncElementRendered();

  let postalCodeInput = form.form.querySelector("#postal-code");
  let addressLevel1Input = form.form.querySelector("#address-level1");
  ok(!postalCodeInput.value, "postal-code should be empty by default");
  ok(!addressLevel1Input.value, "address-level1 should be empty by default");
  ok(postalCodeInput.checkValidity(),
     "postal-code should be valid by default when it is not visible");
  ok(addressLevel1Input.checkValidity(),
     "address-level1 should be valid by default when it is not visible");

  form.remove();
});
</script>

</body>
</html>
